+++ test/sql/other.sql
@@ -29,7 +29,7 @@ values (dlvalue('https://www.debian.org','URL','Sample HTTPS datalink'));
 insert into sample_datalinks2 (link)
 values (dlvalue('http://blah','URL','Broken datalink'));
 
-select (link).url, (link).token is not null as has_token from sample_datalinks2;
+select dlurlcomplete(link), (link)->>'token' is not null as has_token from sample_datalinks2;
 
 create table sample_datalinks3
     as