+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Version 0.2alpha
+-----------
+- datalink type move to pg_catalog
+- internally, datalink is now domain over jsonb
+
 Version 0.1
 -----------
 - initial github version